% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesHorseshoe.R
\name{bayesHorseshoe}
\alias{bayesHorseshoe}
\title{Function to perform Bayesian horseshoe regularization in linear regression}
\usage{
bayesHorseshoe(x, y, T = 10000, B = 5000, thin = 11)
}
\arguments{
\item{x}{Matrix of predictors, size n by p. Should be centered and scaled before calling the function.}

\item{y}{Matrix of response, size p by 1. Should be centered before calling the function.}

\item{B}{Burn-in iterations for Gibbs sampler. Here burn-in iterations are not counted as part of Total iterations.}

\item{thin}{Thinning parameter}

\item{Total}{Total number of iterations for Gibbs sampler that should be stored after thinning}
}
\value{
Returns a matrix of beta coefficients after discarding for burn-in iterations, size Total*p.
}
\description{
Implementation of bhs MATLAB function.
}
\references{
E. Makalic and D. F. Schmidt. (2016) A Simple Sampler for the Horseshoe Estimator.
IEEE Signal Processing Letters, 23(1), 179-182. Jan. 2016, doi: 10.1109/LSP.2015.2503725

Rue, H. (2001). Fast Sampling of Gaussian Markov Random Fields. Journal of the Royal Statistical Society.
Series B (Statistical Methodology), 63(2), 325–338. http://www.jstor.org/stable/2680602

Bhattacharya A., Antik Chakraborty A. & Mallick B. K. (2016). Fast sampling with Gaussian scale mixture priors in high-dimensional regression.
Biometrika, 103(4), 985–991. https://doi.org/10.1093/bi omet/asw042
}
