% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesLassoGamma.R
\name{bayesLassoGamma}
\alias{bayesLassoGamma}
\title{bayesLassoGamma}
\usage{
bayesLassoGamma(x, y, Total = 10000, B = 5000)
}
\arguments{
\item{x}{Matrix of predictors, size n by p}

\item{y}{Matrix of response, size p by 1}

\item{Total}{Total number of iterations for Gibbs sampler}

\item{B}{Burn-in iterations for Gibbs sampler}
}
\value{
Returns a matrix of beta coefficients from Gibbs sampler runs, size Total by p
}
\description{

}
\examples{
x <- matrix(0, nrow=60, ncol=7)
e <- numeric(60)
sigma <- 2.5
for (i in 1:60){
 x[i,] <- stats::rnorm(7, 0, 1)
 e[i] <- stats::rnorm(1,0,sd=sigma^2)
}
y <- numeric(60)
for (i in 1:60){
 y[i] <- 3*x[i,4]+4*x[i,5]+e[i]
}
y <- y - mean(y) # centered y
bayesLassoGamma(x,y) # we expect only fourth and fifth coefficient to be far from zero
}
\references{

}
