% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBayesGammaLasso.R
\name{plotBayesGammaLasso}
\alias{plotBayesGammaLasso}
\title{plotBayesGammaLasso}
\usage{
plotBayesGammaLasso(x, y, Total = 10000, B = 5000)
}
\arguments{
\item{x}{Matrix of predictors, size n by p}

\item{y}{Matrix of response, size p by 1}

\item{Total}{Total number of iterations for Gibbs sampler}

\item{B}{Burn-in iterations for Gibbs sampler}
}
\value{

}
\description{

}
\references{

}
