% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversegaussian.R
\name{inversegaussian}
\alias{inversegaussian}
\title{inversegaussian}
\usage{
inversegaussian(mean, shape)
}
\arguments{
\item{mean}{A vector of mean parameters}

\item{shape}{Single parameter}
}
\value{
A random variable from inverse gaussian distribution
}
\description{
Intakes a vector of means and a single shape parameters.
Generates a vector of random variables from inverse gaussian distribution
 from respective mean and same shape parameters.
}
\examples{
inversegaussian(60,30)
inversegaussian(c(35,20),5)
}
\references{
Michael, John R.; Schucany, William R.; Haas, Roy W. (1976),
"Generating Random Variates Using Transformations with Multiple Roots",
The American Statistician, 30 (2): 88–90, doi:10.1080/00031305.1976.10479147, JSTOR 2683801
}
