% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotB.R
\name{plotB}
\alias{plotB}
\title{plotB}
\usage{
plotB(
  x,
  y,
  Total = 10000,
  B = 5000,
  title = "Plot",
  color = "red",
  includeCI = TRUE,
  includeMAP = TRUE,
  includeLSE = TRUE
)
}
\arguments{
\item{x}{Matrix of predictors, size n by p}

\item{y}{Matrix of response, size p by 1}

\item{Total}{Total number of iterations for Gibbs sampler}

\item{B}{Burn-in iterations for Gibbs sampler}
}
\value{

}
\description{
Generates plots of least squares estimate, maximum-a-posteriori estimate and 95 percent confidence intervals
of beta coefficients for each predictor. Maximum-a-posteriori estimate and 95 percent confidence intervals are generated using
the output from bayesLassoGamma function.
}
\examples{
x <- matrix(0, nrow=60, ncol=7)
e <- numeric(60)
sigma <- 2.5
for (i in 1:60){
 x[i,] <- stats::rnorm(7, 0, 1)
 e[i] <- stats::rnorm(1,0,sd=sigma^2)
}
y <- numeric(60)
for (i in 1:60){
 y[i] <- 3*x[i,4]+4*x[i,5]+e[i]
}
y <- y - mean(y) # centered y
plotB(x,y,color="blue",title="Lasso with Gamma hyperprior",includeMAP=FALSE,includeCI=TRUE,includeLSE =FALSE) #CI
plot(x,y,color="red",title="Lasso with Gamma hyperprior",includeMAP=TRUE,includeCI=TRUE,includeLSE =TRUE) #CI,MAP,LSE

}
\references{

}
